git-svn-id: https://svn.grrrr.org/ext/trunk@867 4d9ac71a-51e6-0310-8455-cad1006bcd31

+++ source/flattr.cpp
@@ -120,7 +120,7 @@ void flext_base::ListAttrib(AtomList &la) const
 		}
 	}
 
-	la(list[0].size()+list[1].size());
+	la((int)(list[0].size()+list[1].size()));
 	int ix = 0;
 	AttrList::iterator it;
 	for(i = 0; i <= 1; ++i)
